refactor: explicit return type
